Add support for native-sparse-cpp's wider 64-bit nnz-offset

+++ jni/src/common.h
@@ -15,7 +15,7 @@ namespace neural_search_jni {
  * Transfer sparse vector data from Java on-heap arrays into off-heap std::vectors.
  *
  * memoryAddresses is a 3-element array of pointers to std::vectors:
- *   [0] -> std::vector<int32_t>*  (indices / CSR indptr)
+ *   [0] -> std::vector<int64_t>*  (indices / CSR indptr — 64-bit cumulative nnz offset)
  *   [1] -> std::vector<int32_t>*  (tokens)
  *   [2] -> std::vector<float>*    (weights)
  *
@@ -36,12 +36,15 @@ inline void transferVectors(
     // an existing vector we must (a) skip the redundant leading 0 and
     // (b) offset every value by the current cumulative nnz so the indptr
     // remains a valid, monotonically-increasing CSR array.
-    std::vector<int32_t>* indicesVec;
+    // CSR indptr is the cumulative-nnz offset (nsparse offset_t = int64_t). Each Java flush sends a
+    // RELATIVE int32 indptr starting at 0 (one flush's nnz stays < INT32_MAX), but the accumulated
+    // offset across flushes can exceed INT32_MAX, so it is stored 64-bit here to match offset_t.
+    std::vector<int64_t>* indicesVec;
     if (memoryAddresses[0] == 0) {
-        indicesVec = new std::vector<int32_t>();
+        indicesVec = new std::vector<int64_t>();
         memoryAddresses[0] = reinterpret_cast<int64_t>(indicesVec);
     } else {
-        indicesVec = reinterpret_cast<std::vector<int32_t>*>(memoryAddresses[0]);
+        indicesVec = reinterpret_cast<std::vector<int64_t>*>(memoryAddresses[0]);
     }
     // Emptiness, not a null address, decides which branch applies: a first call
     // carrying no indptr at all leaves an allocated-but-empty vector behind, and
@@ -50,10 +53,11 @@ inline void transferVectors(
     if (indicesVec->empty()) {
         indicesVec->insert(indicesVec->end(), indices, indices + indicesLen);
     } else {
-        const int32_t offset = indicesVec->back();
-        // Skip indices[0] (always 0) and offset the rest
+        const int64_t offset = indicesVec->back();
+        // Skip indices[0] (always 0) and offset the rest; widen each int32 relative
+        // value to int64 before adding the (possibly >INT32_MAX) cumulative offset.
         for (int i = 1; i < indicesLen; ++i) {
-            indicesVec->push_back(indices[i] + offset);
+            indicesVec->push_back(static_cast<int64_t>(indices[i]) + offset);
         }
     }
 
@@ -87,7 +91,7 @@ inline void transferVectors(
  * a second call a no-op -- closing a buffer twice must not double free.
  */
 inline void freeVectors(int64_t* memoryAddresses) {
-    delete reinterpret_cast<std::vector<int32_t>*>(memoryAddresses[0]);
+    delete reinterpret_cast<std::vector<int64_t>*>(memoryAddresses[0]);
     memoryAddresses[0] = 0;
     delete reinterpret_cast<std::vector<int32_t>*>(memoryAddresses[1]);
     memoryAddresses[1] = 0;
